LPD-93214 buildRest
